+++ src/constants.js
@@ -1,4 +1,4 @@
-export const VERSION = "1.1.2";
+export const VERSION = "1.2.0";
 
 export const DEFAULT_CONFIG = {
   $schema: